+++ PartitionedSolvers/CHANGELOG.md
@@ -6,6 +6,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.1] - 2024-11-11
+
+### Added
+
+- Function `default_solver`
+- First draft of a Newton-Raphson solver in function `newton_raphson`.
+
 ## [0.3.0] - 2024-10-29
 
 ### Changed